+++ test/appifi/lib/samba/model.json
@@ -17,7 +17,31 @@
       "unixPassword": "somestr",
       "smbPassword": "066DDFD4EF0E9CD7C256FE77191EF43C",
       "unixuid": 2000,
-      "unixname": "pandaa",
+      "unixname": "panda",
+      "friends": [],
+      "credentials": {
+        "publicKey": "some str",
+        "privateKey": "some str"
+      },
+      "lastChangeTime": 1489644639494
+    },
+    {
+      "type": "local",
+      "uuid": "6da92303-33a1-4f79-8d8f-a7b6becde6c3",
+      "username": "pandb",
+      "nologin": false,
+      "avatar": null,
+      "email": null,
+      "isFirstUser": true,
+      "isAdmin": true,
+      "home": "b76ed624-d028-4959-b61a-119d212c0cb1",
+      "library": "b76ed624-d028-4959-b61a-119d212c0cb1",
+      "service": "b76ed624-d028-4959-b61a-119d212c0cb1",
+      "password": "$2a$10$ejkgWk/ChHcLyZiic1H2CeLdoVcz15eZ/3Lymj1ncimVvq6E7OvWy",
+      "unixPassword": "somestr",
+      "smbPassword": "066DDFD4EF0E9CD7C256FE77191EF43C",
+      "unixuid": 2001,
+      "unixname": "pandb",
       "friends": [],
       "credentials": {
         "publicKey": "some str",
